fixed type problem

+++ probing/data/dataset.py
@@ -9,7 +9,7 @@ Array = np.ndarray
 class TripletData(torch.utils.data.Dataset):
     def __init__(self, triplets: List[List[int]], n_objects: int):
         super(TripletData, self).__init__()
-        self.triplets = torch.tensor(triplets)
+        self.triplets = torch.tensor(triplets).type(torch.LongTensor)
         self.identity = torch.eye(n_objects)
 
     def encode_as_onehot(self, triplet: Tensor) -> Tensor:
